Search users by username first
I ran into an issue where users with short usernames sometimes don't get returned from the search-users API in the first few rows. This is a problem, as the @ dialog in the comment box uses search-users with a limit of 6, so there is sometimes no way to mention a user, even with an exact username match. This change adds a query to search for an exact match username and present that as the first result, if present.